Moved get_deps_mock() to test case.
